+++ build.gradle
@@ -43,7 +43,7 @@ test {
 }
 
 application {
-    mainClass.set("duke.kbot.KBot")
+    mainClass.set("kbot.main.KBot")
 }
 
 shadowJar {
