+++ src/main/java/kbot/tasks/Event.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package kbot.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
