+++ src/main/java/kbot/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package kbot.tasks;
 
 import java.util.ArrayList;
 
@@ -25,9 +25,9 @@ public class ToDo extends Task {
     /**
      * Constuctor used when we are loading from storage.
      * 
-     * @param name      A String value that states the name of the Task.
+     * @param name        A String value that states the name of the Task.
      * @param isCompleted Boolean of whether Task is completed.
-     * @param tags      A list of tags to identify the task.
+     * @param tags        A list of tags to identify the task.
      */
     public ToDo(String name, boolean isCompleted, ArrayList<String> tags) {
         super(name, TYPE, isCompleted, tags);
