+++ src/main/java/kbot/ui_design/Ui.java
@@ -1,4 +1,4 @@
-package duke.ui_design;
+package kbot.ui_design;
 
 /**
  * Manages all messages that should be printed as the UI component.
